modified ids to int

+++ chatAPI/Models/Chat.cs
@@ -8,7 +8,7 @@ namespace ChatAPI.Models
 {
     public class Chat
     {
-        public long ChatId { get; set; }
+        public int ChatId { get; set; }
         public List<User> Participants { get; set; }
         public List<Message> Messages { get; set;  }
     }
